REF DRY API connection
We were creating connections to the Civis API in multiple places. Centralize everything in a single `open_session` function, and use that everywhere.
