+++ civis/_utils.py
@@ -1,5 +1,15 @@
+from __future__ import absolute_import
+from builtins import super
+
 import re
 import uuid
+
+import requests
+from requests.adapters import HTTPAdapter
+from requests.packages.urllib3.util import Retry
+
+import civis
+from civis import __version__ as client_version
 
 
 UNDERSCORER1 = re.compile(r'(.)([A-Z][a-z]+)')
@@ -20,3 +30,41 @@ def camel_to_snake(word):
 
 def to_camelcase(s):
     return re.sub(r'(^|_)([a-zA-Z])', lambda m: m.group(2).upper(), s)
+
+
+def open_session(api_key, max_retries=5):
+    """Create a new Session which can connect with the Civis API"""
+    civis_version = client_version
+    session = requests.Session()
+    session.auth = (api_key, '')
+    session_agent = session.headers.get('User-Agent', '')
+    user_agent = "civis-python/{} {}".format(civis_version, session_agent)
+    session.headers.update({"User-Agent": user_agent.strip()})
+    max_retries = AggressiveRetry(max_retries, backoff_factor=.75,
+                                  status_forcelist=civis.civis.RETRY_CODES)
+    adapter = HTTPAdapter(max_retries=max_retries)
+    session.mount("https://", adapter)
+
+    return session
+
+
+class AggressiveRetry(Retry):
+    # Subclass Retry so that it retries more things. In particular,
+    # always retry API requests with a Retry-After header, regardless
+    # of the verb.
+    def is_retry(self, method, status_code, has_retry_after=False):
+        """ Is this method/status code retryable? (Based on whitelists and control
+        variables such as the number of total retries to allow, whether to
+        respect the Retry-After header, whether this header is present, and
+        whether the returned status code is on the list of status codes to
+        be retried upon on the presence of the aforementioned header)
+        """
+        if (self.total and
+                self.respect_retry_after_header and
+                has_retry_after and
+                (status_code in self.RETRY_AFTER_STATUS_CODES)):
+            return True
+
+        else:
+            return super().is_retry(method=method, status_code=status_code,
+                                    has_retry_after=has_retry_after)
